Version 0.8.5
Squashed commit of the following:

commit 4e11140875feafb647f3dedc593f2518f835b992
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Fri May 8 11:24:38 2020 -0400

    Fix bug for not warning for nonexistent config files

commit 7327c819b046aaffdd7aacc1a24d13f30ac31e5f
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Thu May 7 19:18:23 2020 -0400

    Version 0.8.5

commit cfe4c88ea2b3fa694e299c307b759e136170292e
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Thu May 7 19:17:09 2020 -0400

    Keep training/validation losses in the evaluations

commit 54810c415711ac79aca08ba5dd24cf78e6e22ff5
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Thu May 7 13:10:15 2020 -0400

    Correct the scaling of hyperparm lambda (no scaling)

commit a88b9e96ba18e4729faa8fb6f9821676221bbe78
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Wed May 6 00:43:02 2020 -0400

    Fix bug of not abling to pickle lambda function

commit 5305262f59c5d55e11f19fcb7932353e8bccadfb
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Tue May 5 23:16:37 2020 -0400

    Learning rate evaluation for pytorch_mlr

commit 96963503ea583b0ede364abcfd38aba3303c2dcc
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Tue May 5 21:51:55 2020 -0400

    Make save/load compatible for SK and PT models

commit b17441d6913d4c99d378040f493924a5b4027e5a
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Tue May 5 19:08:35 2020 -0400

    Adapt klen and lambda eval for pytorch_mlr

commit 2b18b8b08267380fb986a0a9d3e98ec4a619542f
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Tue May 5 15:40:16 2020 -0400

    New config file for pytorch_mlr coverage eval

commit c985e27f26ec6ac5b80e7bea3a282f25a481fc54
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Tue May 5 15:39:16 2020 -0400

    Adapt coverage eval for pytorch_mlr

commit 032f44df84f73cccfebd1e8567d025ddcb897a21
Author: Amine M. Remita <remita.amine@courrier.uqam.ca>
Date:   Mon May 4 23:28:57 2020 -0400

    Add pytorch_mlr and reorganize modules

+++ mlr_kgenomvir/data/build_cv_data.py
@@ -1,7 +1,7 @@
 from .seq_collections import SeqCollection
 from .kmer_collections import GivenKmersCollection
 from .kmer_collections import build_kmers_Xy_data, build_kmers
-from .utils import load_Xy_cv_data, save_Xy_cv_data   
+from ..utils import load_Xy_cv_data, save_Xy_cv_data   
 
 import os.path
 
@@ -32,22 +32,27 @@ def build_cv_data(
     stride = int(fragment_size/fragment_cov)
 
     ## Build data
+    # Set dtype to int32 to comply with pytorch. It is not compatible with
+    # uint64
     if eval_type in ["CC", "FF"]:
 
         if eval_type == "FF":
-            seq_data = seq_data.extract_fragments(fragment_size, stride=stride)
+            seq_data = seq_data.extract_fragments(fragment_size,
+                    stride=stride)
 
             if fragment_count > 1:
                 seq_data = seq_data.stratified_sample(fragment_count)
 
-        X_train, y_train = build_kmers_Xy_data(seq_data, k, full_kmers=full_kmers)
+        X_train, y_train = build_kmers_Xy_data(seq_data, k,
+                full_kmers=full_kmers, dtype=np.int32)
         X_test = "X_train"  # a flag only
         y_test = "y_train"  # a flag only
 
     elif eval_type == "CF":
 
         # Build Train data from complete sequence
-        X_train_kmer = build_kmers(seq_data, k, full_kmers=full_kmers)
+        X_train_kmer = build_kmers(seq_data, k, full_kmers=full_kmers,
+                dtype=np.int32)
 
         X_train = X_train_kmer.data
         y_train = np.asarray(seq_data.labels)
@@ -61,9 +66,9 @@ def build_cv_data(
  
         parents = seq_test.get_parents_rank_list()
 
-        X_test = GivenKmersCollection(seq_test, X_train_kmer_list).data
+        X_test = GivenKmersCollection(seq_test, X_train_kmer_list, 
+                dtype=np.int32).data
         y_test = np.asarray(seq_test.labels)
-
 
     ## Get cross-validation indices
     sss = StratifiedShuffleSplit(n_splits=n_splits, test_size=test_size,
